fix tolerance parameter in rust program

+++ map-converter/src/utils.rs
@@ -42,9 +42,11 @@ pub fn get_source_index_from_target_index(
     get_pixel_index_from_point(&source, source_size.width)
 }
 
-pub fn are_colors_similar(color1: &Color, color2: &Color, squared_tolerance: f32) -> bool {
-    (color1.r - color2.r).pow(2) + (color1.g - color2.g).pow(2) + (color1.b - color2.b).pow(2)
-        <= (squared_tolerance as u8)
+pub fn are_colors_similar(color1: &Color, color2: &Color, squared_tolerance: u16) -> bool {
+    ((color1.r as i32) - (color2.r as i32)).pow(2)
+        + ((color1.g as i32) - (color2.g as i32)).pow(2)
+        + ((color1.b as i32) - (color2.b as i32)).pow(2)
+        <= (squared_tolerance as i32)
 }
 
 pub fn get_adjacent_points(point: &Point, size: &Size) -> [Option<Point>; 4] {
